requirements.txt
slack-bolt==1.21.0
google-api-python-client==2.129.0
google-auth==2.35.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1